chore: util분리 및 경로수정
