+++ lib/utils/playerTableHeader.ts
@@ -1,0 +1,86 @@
+export const recentTableHeader = (position: string) => {
+  if (position === "투수") {
+    return [
+      "일자",
+      "상대",
+      "결과",
+      "점수",
+      "타자",
+      "볼넷",
+      "이닝",
+      "피안타",
+      "피홈런",
+      "사구",
+      "탈삼진",
+      "실점",
+      "자책점",
+    ];
+  } else {
+    return [
+      "일자",
+      "상대",
+      "타율",
+      "타수",
+      "득점",
+      "안타",
+      "2루타",
+      "3루타",
+      "홈런",
+      "타점",
+      "도루",
+      "도실",
+      "볼넷",
+      "사구",
+      "삼진",
+      "병살",
+    ];
+  }
+};
+export const totalTableHeader = (position: string) => {
+  if (position === "투수") {
+    return [
+      "연도",
+      "팀",
+      "평균자책점",
+      "경기",
+      "완투",
+      "완봉",
+      "승",
+      "패",
+      "세",
+      "홀",
+      "승률",
+      "타자",
+      "이닝",
+      "피안타",
+      "피홈런",
+      "볼넷",
+      "사구",
+      "탈삼진",
+      "실점",
+      "자책점",
+    ];
+  } else {
+    return [
+      "연도",
+      "팀",
+      "타율",
+      "경기",
+      "타수",
+      "득점",
+      "안타",
+      "2루타",
+      "3루타",
+      "홈런",
+      "타점",
+      "도루",
+      "도실",
+      "볼넷",
+      "사구",
+      "삼진",
+      "병살",
+      "장타율",
+      "출루율",
+    ];
+  }
+};
